# SR-12021: Improve the performance of NSData Base64 Encoding.