@spevans 2020-01-21
@swift-ci test