@spevans 2020-01-25
@swift-ci test and merge